+++ test/acceptance/tests/hcp/terraform/ap/variables.tf
@@ -70,7 +70,7 @@ variable "consul_image" {
 variable "consul_ecs_image" {
   description = "Consul ECS image to use."
   type        = string
-  default     = "hashicorppreview/consul-ecs:0.7.3"
+  default     = "hashicorp/consul-ecs:0.7.3"
 }
 
 variable "consul_server_address" {